improve state printing for multi-line values
In the current code, if the definition of a state value stretches over more than one line, deparse breaks it up into multiple lines; the unlist in the current code then returns more lines than there are states.   Then you end up with many repeated names, and difficult to interpret output.   This takes care of it, while remaining identical in the case where all state values are defined in a single line.