allow commas in numbers
in data frames being read in.   I'm finding my colleagues who are used to Excel are prone to putting them in.

If it finds commas in a column, it checks whether that column (if you remove the commas) can be converted to numeric without errors.  If so, then it does so.