improve warning in look_up (#285)
* improve warning in look_up

If any NA results are returned, the current version just notes NA results are being returned, which doesn't help find which look_up statement is causing the problem.   New warning message returns the names of the data object being looked up in (since I don't think the name of the data object itself is accessible from within the call - the call has already had the data.frame substituted in -- it's not even present in the full call stack I can get from within the function), as well as the selection arguments passed in.   This makes it reasonably easy to find the offending call.

* slightly improved version