Update strategy_eval.R (#290)
This was an area that many of the modelers who were trying to understand the code found confusing.   I've pulled out most of the expansion code into a separate function so that evaluate_strategy now has a simple flow:   expand if necessary, evaluate the transitions and states, compute the counts, compute the values.