add HIV and THR models from DMHEE book to tests (#296)
* add tests of DMHEE models for HIV and THR to tests

This will ensure that these run each time the package is tested, and are compared directly to the numbers from the Excel files accompanying the book.   I think this will be very useful for persuading regulators (and others) that the package gets correct answers.

* update i-reproduction.Rmd

Change the parameters in the THR model to exactly match those in the Excel sheet accompanying DMHEE book.  (They were very slightly off; they now match the ones in the test.)

* move more exact tests of DMHEE models from test_run_model over to test_dmhee.